requirements.txt
gradio[mcp]
selenium
webdriver-manager
beautifulsoup4
requests
python-dotenv
openai
pillow